CreateUser was added
