+++ createUser.html
@@ -7,7 +7,7 @@
     <title>CreateUser</title>
 </head>
 <body>
-    <script type="module" src="./createUser.js"></script>
+    <script type="module" src="./CreateUser/createUser.js"></script>
     <form id="create-user-form">
         <div class="form-control">
           <label for="post-title">Name:</label>
