--- createUserController.js
@@ -1,3 +1,0 @@
-async function getNewUserData() {
-    
-}
